Refactoring home screen and publish develop branch
