+++ lib/screens/splash_screen.dart
@@ -2,7 +2,7 @@ import 'package:animated_splash_screen/animated_splash_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:lottie/lottie.dart';
 import 'package:notes_app/constants/app_colors.dart';
-import 'package:notes_app/screens/test_splash_icon.dart';
+import 'package:notes_app/screens/home_screen.dart';
 
 class SplashScreen extends StatelessWidget {
   const SplashScreen({super.key});
